cython files

+++ scripts/block_matrix.py
@@ -2,6 +2,8 @@ import argparse
 import logging
 import sys
 
+import numpy as np
+
 from ghw import pklLoad, pklSave
 from ghw.block_matrix import BMat, dflist
 from ghw.db import Db
@@ -16,7 +18,7 @@ def allSenders(dbpath, year):
 
 
 def main(socio_matrix, dbpath, distance, year):
-    dmat = BMat.dmat(socio_matrix.matrix, dflist[f"{distance}_metric"])
+    dmat = BMat.dmat(socio_matrix.matrix.astype(np.float32), dflist[f"{distance}_metric"])
     block_matrix = BMat(socio_matrix, distance, dmat)
     pklSave(args.output, block_matrix)
